+++ src/pipeline/transaction/runTQLMutation.ts
@@ -28,7 +28,7 @@ export const runTQLMutation: PipelineOperation = async (req, res) => {
   if (!mutateTransaction) {
     throw new Error("Can't create transaction");
   }
-  // console.log('tqlRequest', tqlRequest);
+  // console.log('tqlRequest!', JSON.stringify(tqlRequest, null, 2));
   // deletes and pre-update deletes
   const tqlDeletion =
     tqlRequest.deletionMatches &&